requirements.txt

streamlit>=1.37.0
openai>=1.43.0
pandas>=2.2.2
pillow>=10.4.0
pypdf>=4.3.1
PyMuPDF>=1.24.9
python-dotenv>=1.0.1
